# Conflicts:
#	tests/VStore.UnitTests/ImageRendering/MemoryConsumptionTests.cs

+++ src/VStore/ImageRendering/MemoryLimitedException.cs
@@ -4,7 +4,7 @@ namespace NuClear.VStore.ImageRendering
 {
     public sealed class MemoryLimitedException : Exception
     {
-        public MemoryLimitedException(int memoryRequested, int memoryThreshold)
+        public MemoryLimitedException(int memoryRequested, long memoryThreshold)
             : base($"Memory limits applied. Requested memory: '{memoryRequested}', memory threshold: '{memoryThreshold}'.")
         {
         }
